apps/compat: revert float64 workaround to idiomatic int types; comparator now normalizes "integer" ↔ "number"
The float64-everywhere choice in the dynamic batch was the wrong direction
— it let TypeScript's "every number is a 64-bit float" worldview
constrain mcpkit's richer Go type system. JSON Schema "integer" is a
subtype of "number"; any integer value validates against both. mcpkit's
"integer" emission is the more precise description.

Real fix: normalize "integer" → "number" in the drift comparator. Now
fixtures use Go-idiomatic types (int for counts and byte totals, float64
for ratios and percentages). Both sides keep their natural emit
behavior; the comparator treats them as equivalent.

Affected:

- system-monitor: CPU count, byte totals, percent (whole), CPU ticks,
  uptime seconds → int / int64 / uint64
- cohort-heatmap: counts and indexes → int; Retention stays float64
  (ratio)
- customer-segmentation: revenue, employee count, account age, support
  tickets, NPS, count → int; EngagementScore stays float64
- debug-server: counter, delayMs, large input length → int / int64
- integration: unchanged (no numeric fields)

Issue 548's Gap 1 entry is being downgraded — this isn't a mcpkit
schema-gen gap, it's a comparator concern, now resolved.

+++ examples/apps/compat/cohort-heatmap/main.go
@@ -2,10 +2,11 @@
 //
 // One tool — get-cohort-data — with an enum/default-laden input schema and
 // a deeply nested output schema. Defaults are single values (no commas), so
-// struct-tag reflection works cleanly. All numerics use float64 so the
-// auto-derived schema emits "type": "number" matching upstream's zod-from-
-// `z.number()` shape (mcpkit's invopop reflects int → "integer", which
-// diverges from upstream — same gap as system-monitor).
+// struct-tag reflection works cleanly. Idiomatic Go types throughout —
+// counts and indexes are int; Retention is float64 (it's a ratio like
+// 0.85). The drift comparator normalizes "integer" ↔ "number" so the
+// idiomatic types pass cleanly against upstream's "number"-everywhere
+// zod-derived schema.
 //
 // Run:  EXT_APPS_DIR=/tmp/ext-apps PORT=3101 go run .
 package main
@@ -25,24 +26,24 @@ import (
 )
 
 type cohortInput struct {
-	Metric      string  `json:"metric,omitempty" jsonschema:"enum=retention,enum=revenue,enum=active,default=retention"`
-	PeriodType  string  `json:"periodType,omitempty" jsonschema:"enum=monthly,enum=weekly,default=monthly"`
-	CohortCount float64 `json:"cohortCount,omitempty" jsonschema:"minimum=3,maximum=24,default=12"`
-	MaxPeriods  float64 `json:"maxPeriods,omitempty" jsonschema:"minimum=3,maximum=24,default=12"`
+	Metric      string `json:"metric,omitempty" jsonschema:"enum=retention,enum=revenue,enum=active,default=retention"`
+	PeriodType  string `json:"periodType,omitempty" jsonschema:"enum=monthly,enum=weekly,default=monthly"`
+	CohortCount int    `json:"cohortCount,omitempty" jsonschema:"minimum=3,maximum=24,default=12"`
+	MaxPeriods  int    `json:"maxPeriods,omitempty" jsonschema:"minimum=3,maximum=24,default=12"`
 }
 
 type cohortCell struct {
-	CohortIndex   float64 `json:"cohortIndex"`
-	PeriodIndex   float64 `json:"periodIndex"`
+	CohortIndex   int     `json:"cohortIndex"`
+	PeriodIndex   int     `json:"periodIndex"`
 	Retention     float64 `json:"retention"`
-	UsersRetained float64 `json:"usersRetained"`
-	UsersOriginal float64 `json:"usersOriginal"`
+	UsersRetained int     `json:"usersRetained"`
+	UsersOriginal int     `json:"usersOriginal"`
 }
 
 type cohortRow struct {
 	CohortID      string       `json:"cohortId"`
 	CohortLabel   string       `json:"cohortLabel"`
-	OriginalUsers float64      `json:"originalUsers"`
+	OriginalUsers int          `json:"originalUsers"`
 	Cells         []cohortCell `json:"cells"`
 }
